# docs/swagger.yaml
definitions:
  api.LoginBodyDto:
    properties:
      password:
        description: 密码
        type: string
      username:
        description: 用户名
        type: string
    type: object
  api.LoginSuccessResponse:
    properties:
      token:
        type: string
    type: object
  api.RegisterBodyDto:
    properties:
      code:
        description: 邮箱验证码
        type: string
      email:
        description: 邮箱
        type: string
      password:
        description: 密码
        type: string
      username:
        description: 用户名
        type: string
    type: object
  cli.CreateModuleBody:
    properties:
      name:
        description: 模块名称
        type: string
    type: object
  resp.Result:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
info:
  contact: {}
  title: 男生自用 API 接口文档
  version: "1.0"
paths:
  /admin/api/user/list:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: resp
          schema:
            type: string
      summary: 用户列表
      tags:
      - 管理后台-用户
  /api/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/api.LoginBodyDto'
      produces:
      - application/json
      responses:
        "200":
          description: resp
          schema:
            allOf:
            - $ref: '#/definitions/resp.Result'
            - properties:
                data:
                  $ref: '#/definitions/api.LoginSuccessResponse'
              type: object
      summary: 用户登录
      tags:
      - 用户
  /api/auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/api.RegisterBodyDto'
      produces:
      - application/json
      responses:
        "200":
          description: resp
          schema:
            allOf:
            - $ref: '#/definitions/resp.Result'
            - properties:
                data:
                  $ref: '#/definitions/api.LoginSuccessResponse'
              type: object
      summary: 用户注册
      tags:
      - 用户
  /api/cli/module/create:
    post:
      consumes:
      - application/json
      description: get string by ID
      parameters:
      - description: 模块名称
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/cli.CreateModuleBody'
      produces:
      - application/json
      responses:
        "200":
          description: resp
          schema:
            allOf:
            - $ref: '#/definitions/resp.Result'
            - properties:
                data:
                  type: boolean
              type: object
        "500":
          description: resp
          schema:
            allOf:
            - $ref: '#/definitions/resp.Result'
            - properties:
                data:
                  type: string
              type: object
      summary: 创建模块
      tags:
      - cli
swagger: "2.0"